reorganized top menu bar
