+++ src/windowmanagement/NewWindowMenu.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Menu } from '@mantine/core';
 
-import { ChevronRightIcon, PanelIcon } from '@/icons/icons';
 import { IconSize } from '@/types/enums';
 
 import { menuItemsDB } from './data/MenuItems';
@@ -12,40 +11,23 @@ interface NewWindowMenuProps {
 }
 export function NewWindowMenu({ addWindow }: NewWindowMenuProps) {
   return (
-    <Menu
-      position={'right-start'}
-      withinPortal={false}
-      closeOnItemClick={false}
-      menuItemTabIndex={0}
-      withArrow
-      arrowPosition={'center'}
-    >
-      <Menu.Target>
+    <>
+      <Menu.Label>Add New Window</Menu.Label>
+      {menuItemsDB.map((item) => (
         <Menu.Item
-          leftSection={<PanelIcon />}
-          rightSection={<ChevronRightIcon size={IconSize.sm} />}
+          key={item.componentID}
+          leftSection={item.icon?.(IconSize.xs)}
+          onClick={() => {
+            addWindow(item.content, {
+              title: item.title,
+              position: item.preferredPosition,
+              id: item.componentID
+            });
+          }}
         >
-          Windows
+          {item.title}
         </Menu.Item>
-      </Menu.Target>
-      <Menu.Dropdown>
-        <Menu.Label>Add New Window</Menu.Label>
-        {menuItemsDB.map((item) => (
-          <Menu.Item
-            key={item.componentID}
-            leftSection={item.icon?.(IconSize.xs)}
-            onClick={() => {
-              addWindow(item.content, {
-                title: item.title,
-                position: item.preferredPosition,
-                id: item.componentID
-              });
-            }}
-          >
-            {item.title}
-          </Menu.Item>
-        ))}
-      </Menu.Dropdown>
-    </Menu>
+      ))}
+    </>
   );
 }
